tg_bot/backend/app/requirements.txt
fastapi
uvicorn
aiogram
tortoise-orm
requests